Adding public Wait for scheduler

+++ scheduler/job.go
@@ -22,20 +22,22 @@ type job struct {
 	id       string
 	cmd      *exec.Cmd
 	outputSt *stream.Stream
-	done     chan bool
 	sts      *JobStatus
 	m        logsync.Mutex
+	wg       *sync.WaitGroup
 }
 
 // newJob creates a new job
-func newJob() *job {
+func newJob(wg *sync.WaitGroup) *job {
+	wg.Add(1)
+
 	id := generateRandomId()
 	p := &job{
 		id:       id,
-		done:     make(chan bool),
 		outputSt: stream.New(),
 		sts:      &JobStatus{Type: Idle, ExitCode: -1},
 		m:        logsync.New(fmt.Sprintf("job %s", id)),
+		wg:       wg,
 	}
 
 	return p
@@ -63,6 +65,8 @@ func (j *job) start(executable string, args ...string) error {
 	log.Debugf("Starting: %s\n", j.cmd.Path)
 
 	go func() {
+		defer j.wg.Done()
+
 		err := j.run(errCh)
 		if err != nil {
 			j.updateStatus(Errored)
@@ -79,6 +83,10 @@ func (j *job) start(executable string, args ...string) error {
 // stop stops a running process
 func (j *job) stop() error {
 	j.m.WLock("stop")
+	if j.cmd == nil {
+		j.m.WUnlock("stop")
+		return fmt.Errorf("job not started")
+	}
 	err := j.cmd.Process.Kill()
 	j.m.WUnlock("stop")
 
@@ -105,11 +113,6 @@ func (j *job) status() *JobStatus {
 
 	// JobStatus needs to be cloned or a race condition might happen (since it's a reference)
 	return j.sts.clone()
-}
-
-// wait blocks until the process is completed
-func (j *job) wait() {
-	<-j.done
 }
 
 func (j *job) run(started chan error) error {
@@ -158,8 +161,6 @@ func (j *job) run(started chan error) error {
 		j.updateStatus(Exited)
 	}
 
-	j.done <- true
-
 	return nil
 }
 
@@ -207,6 +208,8 @@ func (j *job) updateStatus(st StatusType) {
 	defer j.m.WUnlock("updateStatus")
 
 	switch j.sts.Type {
+	case Exited | Errored | Killed:
+		// Do not update the status, the previous one is the one we want to keep
 	case Running:
 		j.sts.Type = st
 		j.outputSt.Close()
